clock at the end of the day is 23:59:59.999
Ensure that the test fixtures will return; the previous changes
correctly forward the time to end of day across zones and
datelines.  One side effect  is that the date formatted for the end
of the day uses all the microseconds for the day as well.  This
changes the query from the form of 23:59:59.000Z to 23:59:59.999Z.
This also ensures that anything that happens during that one second
will be included rather than excluded.