allow client to load successfully without requiring API SECRET
This patch allows the operator to control whether or not the API Secret
authentication dialog will display by default or not.  Historically Nightscout
allows a configuration to be readable and to send alarm information to all
displays.
A change in in dev in February to support authentication and api v3 introduced
a new configuration where websocket communication is authenticated, and the API
Secret modal will display if it authenticates in error.  The side-effect is
that all new instances, even readable instances with glucose data, can't be
seen or monitored without putting in the API Secreet or explicitly dismissing
it.
This patch returns to the default behavior while allowing to select the new one
using authenticationPromptOnLoad in the settings communicated from server to
client. If people strongly prefer, we can reverse to the new behavior over
time.
To select the old behavior with no API Secret needed for readabable data, do nothing.
To select the new behavior where the API Secret prompts occur on load, set
`AUTHENTICATION_PROMPT_ON_LOAD=1`.