add notes regarding handling unauthorized ACK request
When someone is looking at Nightscout and needs the alarm silenced, it is very
desirable to always silence the local UI. This patch documents some of the
working code around handling the alarm notification process, as well as
provides commentary on handling unauthorized scenarios.  There are some open
questions such as how to update the permission set after authorization.